Changed the city to the number of friends and put it in a separate component. On the search page, added the number of friends and date of birth
