+++ src/assets/styles/screens/Profile.module.scss
@@ -13,31 +13,6 @@
     align-items: center;
     justify-content: space-between;
     margin-bottom: 20px;
-
-    .title {
-      font-size: 16px;
-      position: relative;
-      left: 100px;
-      bottom: 0px;
-      color: #fefefe;
-      display: flex;
-      flex-direction: column;
-    }
-
-    .info {
-      font-size: 13px;
-      position: relative;
-      left: 0;
-      bottom: 0;
-      color: #8a8a8a;
-      margin-top: 5px;
-    }
-
-    .button {
-      position: relative;
-      right: 0;
-    }
-
   }
 
   .postsContainer {
