+++ src/components/ui/post/Post.module.scss
@@ -12,13 +12,13 @@
 
 
     .content {
-      margin: 20px 0;
+      margin: 12px 0;
 
       img {
         width: 100%;
         height: 100%;
         object-fit: cover;
-        margin-top: 20px;
+        margin-top: 12px;
         border-radius: 10px;
       }
     }
