+++ src/components/ui/profile-info/ProfileInfo.module.scss
@@ -1,0 +1,20 @@
+
+.title {
+  font-size: 16px;
+  position: relative;
+  bottom: 0;
+  color: #fefefe;
+  display: flex;
+  flex-direction: column;
+}
+
+.info {
+  font-size: 13px;
+  position: relative;
+  left: 0;
+  bottom: 0;
+  color: #8a8a8a;
+  margin-top: 5px;
+}
+
+
