+++ src/components/ui/profile-info/ProfileInfo.tsx
@@ -1,24 +1,23 @@
-import styles from '@/assets/styles/screens/Profile.module.scss'
-import Image from 'next/image'
-import home from '@/assets/img/home.svg'
-import Button from '@/components/ui/button/Button'
-import { useCreateConversation } from '@/hooks/useCreateConversation'
+import styles from './ProfileInfo.module.scss'
+
 import { IUser } from '@/types/user.interface'
-import { getFriendAction } from '@/lib/get-friend-action'
 import { useDateWithYear } from '@/hooks/useDate'
-import { FriendActions } from '@/components/screens/friends/friend-actions/FriendActions'
+import { FriendActions } from '@/components/ui/friend-actions/FriendActions'
 import React from 'react'
 import { getQuantityFriends } from '@/lib/get-quantity-friends'
-import { useUserProfile } from '@/hooks/useProfile'
+import Link from 'next/link'
 
 interface Props {
 	profileId: string
 	userProfile: IUser | undefined
-	isLoading: boolean
-	refetchUserProfile: any
+	isLoading?: boolean
+	refetchUserProfile?: () => void
+	style?: any
 }
 
-export const ProfileInfo: React.FC<Props> = ({ profileId, userProfile, isLoading, refetchUserProfile }) => {
+export const ProfileInfo: React.FC<Props> = ({ style, profileId, userProfile, isLoading, refetchUserProfile }) => {
+
+
 
 	const userFullName = isLoading
 		? 'SKELETON'
@@ -27,17 +26,19 @@ export const ProfileInfo: React.FC<Props> = ({ profileId, userProfile, isLoading
 	const quantityFriends = getQuantityFriends(userProfile?.friends.length)
 
 	return (
-		<div className={styles.uInfo}>
-			<div className={styles.title}>
-				<span>{userFullName}</span>
+		<>
+			<div className={styles.title} style={style}>
+				<Link href={`/profile/${profileId}`}>
+					<span>{userFullName}</span>
+				</Link>
 				<span className={styles.info}>
 						{quantityFriends}
-					</span>
+				</span>
 				<span className={styles.info}>
 						{useDateWithYear(userProfile?.createdAt) + 'г.'}
-					</span>
+				</span>
 			</div>
-			<FriendActions refetchUserProfile={refetchUserProfile} friendId={profileId} />
-		</div>
+		</>
+
 	)
 }
